requirements.txt
click>=8.1.0
rich>=13.0.0
pyyaml>=6.0
jinja2>=3.1.0
requests>=2.28.0
kubernetes>=27.0.0
ruamel.yaml>=0.17.0
